requirements.txt
ultralytics
torch
opencv-python
numpy
pandas
matplotlib
gradi
